third_party/blink/web_tests/external/wpt/css/cssom/caretRangeFromPoint-replace-document.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] document.caretRangeFromPoint(0, 0)
  assert_equals: expected 0 but got 1
Harness: the test ran to completion.
